camera: move Camera_Apply to TRX
This moves Camera_Apply and the linked private functions to ensure the
environment is correctly setup to TRX.
