+++ src/libtrx/include/libtrx/game/camera/common.h
@@ -4,7 +4,7 @@
 #include "../rooms/types.h"
 
 extern void Camera_Update(void);
-extern void Camera_Apply(void);
+void Camera_Apply(void);
 void Camera_Move(const GAME_VECTOR *target, int32_t speed);
 bool Camera_IsChunky(void);
 void Camera_SetChunky(bool is_chunky);
