+++ src/libtrx/include/libtrx/game/sound/common.h
@@ -27,4 +27,5 @@ extern void Sound_StopAmbientSounds(void);
 extern void Sound_StopAll(void);
 extern bool Sound_Effect(
     SOUND_EFFECT_ID sfx_num, const XYZ_32 *pos, uint32_t flags);
+extern void Sound_StopEffect(SOUND_EFFECT_ID sfx_num);
 bool Sound_IsAvailable(SOUND_EFFECT_ID sfx_num);
